==> 17 Pandas_Essential/02_Joining Multiple Data Tables.ipynb <==
import pandas as pd
import os

master = pd.read_pickle(os.path.join(".", "pickle_files","master.pickle"))
scoring = pd.read_pickle(os.path.join(".", "pickle_files","scoring.pickle"))
teams = pd.read_pickle(os.path.join(".", "pickle_files","teams.pickle"))
team_splits = pd.read_pickle(os.path.join(".","pickle_files", "team_splits.pickle"))

master.head(2)

scoring.head(2)

pd.merge(master, scoring, left_index=True, right_on="playerID").head()

scoring.index

scoring.index + 3

scoring.index = scoring.index + 3

pd.merge(master, scoring, left_index=True, right_on="playerID").head()

pd.merge(master, scoring.set_index("playerID", drop=True),
                                   left_index=True, right_index=True).head()

scoring = scoring.reset_index(drop=True)
pd.merge(master, scoring, left_index=True,
                  right_on="playerID").head()


print(

    pd.merge(master, scoring, left_index=True,
             right_on="playerID").shape,
    
    pd.merge(master, scoring, left_index=True,
                  right_on="playerID", how="right").shape
)


# Drop random records
master2 = master.drop(master.sample(5).index)

print(

    pd.merge(master2, scoring, left_index=True,
             right_on="playerID").shape,
    
    pd.merge(master2, scoring, left_index=True,
                  right_on="playerID", how="right").shape
)


merged = pd.merge(master2, scoring, left_index=True,
                  right_on="playerID", how="right", indicator=True)
merged.head()

merged["_merge"].value_counts()

merged[merged["_merge"] == "right_only"].head()

scoring2 = scoring.drop(scoring.sample(1000).index)

merged = pd.merge(master2, scoring2, left_index=True,
                  right_on="playerID",
                  how="outer", indicator=True)

merged[(merged["_merge"] == "left_only") |
       (merged["_merge"] == "right_only")
      ].sample(3)

# https://pandas.pydata.org/pandas-docs/stable/api.html#string-handling

merged[merged["_merge"].str.endswith("only")].sample(5)

pd.merge(master, scoring,
                 left_index=True, right_on="playerID",
                 validate="1:m").head()

try:
    pd.merge(master, scoring,
                 left_index=True, right_on="playerID",
                 validate="1:1")
except Exception as e:
    print(e)

merged = pd.merge(master, scoring,
                 left_index=True, right_on="playerID")
merged.head()

merged = merged.filter(regex="^(?!(birth)).*")
merged.head()

merged.to_pickle(os.path.join(".","pickle_files", "scoring.pickle"))

teams.sample(2)

team_splits.sample(2)

teams2 = teams[["tmID", "name"]]
teams2 = teams2.drop_duplicates()
teams2["tmID"].value_counts().head()

teams2.loc[teams2["tmID"] == "CHI"]

teams2 = teams2[teams2["tmID"] == "CHI"]
team_splits2 = team_splits[team_splits["tmID"] == "CHI"].sample(2)
team_splits2

pd.merge(teams2, team_splits2)

pd.merge(team_splits, teams, 
                        left_on=["tmID", "year"],
                        right_on=["tmID", "year"]).head(4)

try:
    pd.merge(team_splits, teams, 
                        left_on=["tmID", "year"],
                        right_on=["year", "tmID"]).head()
except Exception as e:
    print(e)

pd.merge(team_splits, teams, 
                        on=["tmID", "year"]).head(4)

pd.merge(team_splits, teams, 
                        on=["tmID"]).head(4)

pd.merge(team_splits, teams, 
                        on=["tmID"]).head().filter(like="year")

pd.merge(team_splits, teams, 
                        on=["tmID"], 
         suffixes=["_team_splits", "_teams"]).head().filter(like="year")

# Like means: "year" in x == True

merged = pd.merge(team_splits, teams, on=["tmID", "year"])
merged.head(3)

merged.to_pickle(os.path.join(".","pickle_files", "team_splits.pickle"))

# Bonus

# This is to show join going bad

df1 = pd.DataFrame([
        ["hopkide01", 4, "R"],
        ["mairad01", 1, "C"],
        ["chaseke01", 5, "R"]
    ],
    columns=["playerID", "goals", "position"]
)

df2 = pd.DataFrame([
        ["hopkide01", "Dean Hopkins"],
        ["mairad01", "Adam Mair"],
        ["chaseke01", "Kelly Chase"]
    ],
    columns=["playerID", "name"]
)    

pd.merge(df1, df2)

# This is to show join going bad

df1 = pd.DataFrame([
        ["hopkide01", 4, "R"],
        ["mairad01", 1, "C"],
        ["chaseke01", 5, "R"]
    ],
    columns=["playerID", "goals", "position"]
)

df2 = pd.DataFrame([
        ["hopkide01", "Dean Hopkins", "Right Wing"],
        ["mairad01", "Adam Mair", "Centre"],
        ["chaseke01", "Kelly Chase", "Right Wing"]
    ],
    columns=["playerID", "name", "position"]
)

pd.merge(df1, df2)

# df.join

df1 = pd.DataFrame([
        ["hopkide01", 4],
        ["mairad01", 1],
        ["chaseke01", 5]
    ],
    columns=["playerID", "goals"]
)


df1 = df1.set_index("playerID", drop=True)

df2 = pd.DataFrame([
        ["hopkide01", "R"],
        ["mairad01", "C"],
        ["chaseke01", "R"]
    ],
    columns=["playerID", "position"],
)
df2 = df2.set_index("playerID", drop=True)

df3 = pd.DataFrame([
        ["hopkide01", "Dean Hopkins"],
        ["mairad01", "Adam Mair"],
        ["chaseke01", "Kelly Chase"]
    ],
    columns=["playerID", "name"]
)
df3 = df3.set_index("playerID", drop=True)

df1.join([df2, df3])
