==> 17 Pandas_Essential/05_Timeseries.ipynb <==
import pandas as pd
import os

master = pd.read_pickle(os.path.join(".","pickle_files", "master.pickle"))
team_splits = pd.read_pickle(os.path.join(".", "pickle_files", "team_splits.pickle"))

master.head()

master = master.assign(birthDate= pd.to_datetime({"year": master.birthYear, "month":master.birthMon, "day":master.birthDay}))

master.drop(columns=["birthYear", "birthMon", "birthDay"], inplace=True)

master.head()

ts = master.birthDate[0]
ts

print(u"{} , {} {} {}".format(ts.day_name(), ts.day, ts.month_name(), ts.year) )

tsz = ts.tz_localize("Europe/London")
tsz

tsz.tz_convert("America/Toronto")

years = master.birthDate.dt.year
years.head()

strings = master.birthDate.dt.strftime("%Y-%m-%d-%a")
strings.head()

strings = master.birthDate.dt.strftime("%Y-%m-%d")
strings.head()

dates = pd.to_datetime(strings)
dates.head()

strings = strings.str.replace("-", "xx")
strings.head()

try: 
    pd.to_datetime(strings)
except Exception as e:
    print(e)

parsed = pd.to_datetime(strings, format="%Yxx%mxx%d")
parsed.head()

string_error = strings.copy()
string_error.iloc[3] = "xx23"
try: 
    pd.to_datetime(string_error, format="%Yxx%mxx%d")
except Exception as e:
    print(e)

parsed = pd.to_datetime(string_error, format="%Yxx%mxx%d", errors="coerce")
parsed.head()
# NaT stands for Not A Time

Don't use errors="ignore" it ignores the erros and gives unpredictable results

parsed = pd.to_datetime(string_error, format="%Yxx%mxx%d", errors="ignore")
parsed.head()
# NaT stands for Not A Time

birth_dates = pd.Series(master.index, index=master.birthDate)
birth_dates.head()

birth_dates.index

**Always sort the index for time series analysis**

birth_dates = birth_dates.sort_index()
birth_dates.head()

**Very tricky one Keep this in mind**

print(birth_dates[1980])  # this will give the index position
print(birth_dates.iloc[1980])  # this will give the index position
born_1980 = birth_dates["1980"]  # this will get all players born in 1980
born_1980.head()

birth_dates["1980-07"].head()

birth_dates["1980-07-24"].head()

birth_dates["1989-07-20" : "1989-08-15"].head()  # both upper and lower limits are inclusive

# Timedelta

diff = birth_dates.index[1] - birth_dates.index[0]
diff

birth_dates.index[2]

birth_dates.index[2] + diff

birth_dates.index - diff

diff.components

diff = diff + "5H" - "10M"
diff

diff.round("10D")

diff.round("3H")

diff.round("1H")

birth_dates.index
# check the dytpe of the output

birth_dates=birth_dates.to_period(freq="D")
birth_dates.index
# check the dytpe of the output

birth_dates.index[0]

start = birth_dates.index[0].start_time
print(start)
end = birth_dates.index[0].end_time
print(end)

start < (start + pd.Timedelta("5H")) < end

start < (start + pd.Timedelta("49H")) < end

birth_dates.index.year

**Remember birth_dates[1940] is location not the year. birth_dates["1940"] is year**

birth_dates["1940"]

birth_dates["1940-03-22"]

# Advanced Strategies

resample is kind of group by for time series data

birth_dates.resample("1M")

birth_dates.resample("1M").count().head()

birth_dates.resample("2M").count().head()

birth_dates.resample("1Q").count().head()

birth_dates.resample("1W").count().head()

from pandas.tseries.offsets import BDay
p = birth_dates.index[2]
print(p.to_timestamp())                  # We are converting this to time stamp because Bday doesn't work well with Period Index
print(p.to_timestamp() + BDay(9)) 
# this will add 9 business days 


birth_dates.head()

# shift freq should always match with index freq
birth_dates.shift(1, freq="D").head()

try:
    birth_dates.shift(1, freq="M").head()
except Exception as e:
    print(e)
# you have to modify the index to period index to Month

# this will move the row
birth_dates.shift(1).head()

team_splits.head()

team_splits.drop(columns="tmID", inplace=True)


team_splits=team_splits.set_index(["name", "year"])
team_splits.head()

month = team_splits.columns.map(lambda x: x[:3])
metric = team_splits.columns.map(lambda x: x[3:])
print(month)
print(metric)

midx = pd.MultiIndex.from_arrays([month, metric])
midx

team_splits.columns = midx
team_splits.head()

team_splits.columns.levels[0].name= "month"
team_splits.columns.levels[1].name= "metric"


team_splits.head()

team_splits = team_splits.stack(level="month")
team_splits.head()

team_splits=team_splits.reset_index()
team_splits.head()

starting_date = team_splits.apply(lambda x: pd.Period("{}-{}".format(x.year, x.month)), axis=1) # axix 1 for row
starting_date.head()

starting_date[1]

team_splits.drop(columns="year", inplace=True)
team_splits = team_splits.assign(month = starting_date)
team_splits.head()

team_splits = team_splits.set_index(["month"])
team_splits.head()

team_splits.columns.name= None

team_splits.head()

team_splits.groupby(["name"]).resample("Q").mean().head()

quarter_means = team_splits.groupby(["name"]).resample("Q").mean().head()
quarter_means.head()

team_splits.to_pickle(os.path.join(".", "pickle_files", "team_splits_period.pickle"))
birth_dates.to_pickle(os.path.join(".", "pickle_files", "birth_dates.pickle"))
